move main from CheckBitstreamIterator to app.checker.ChecksumWorker
